# docs/swagger.yaml
definitions:
  attachment.Attachment:
    properties:
      createdAt:
        type: string
      height:
        type: integer
      id:
        type: string
      mime:
        type: string
      serverId:
        type: string
      size:
        type: integer
      thumbUrl:
        type: string
      url:
        type: string
      userId:
        type: string
      width:
        type: integer
    type: object
...
info:
  contact: {}
  description: Armonic API backend server.
  title: Armonic API
  version: "1.0"
paths:
  /attachment/{id}:
    get:
      description: Member-only. Serves the sanitized image bytes.
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: attachment not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Fetch an attachment
      tags:
      - Attachment
...
  /me/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image through the same pipeline as attachments and points
        the caller's avatar at it.
      parameters:
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/attachment.Attachment'
        "413":
          description: file exceeds the maximum upload size
          schema:
            type: string
        "415":
          description: unsupported image format
          schema:
            type: string
        "429":
          description: rate limit exceeded
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Set own avatar
      tags:
      - Users
...
  /server/{id}/upload:
    post:
      consumes:
      - multipart/form-data
      description: Member-only. Multipart upload of one image (field "file"). The
        file is validated by magic number, decoded, stripped of all metadata by re-encoding
        its pixels, thumbnailed and stored. The uploaded bytes are discarded.
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/attachment.Attachment'
        "400":
          description: image could not be decoded
          schema:
            type: string
        "413":
          description: file exceeds the maximum upload size
          schema:
            type: string
        "415":
          description: unsupported image format
          schema:
            type: string
        "422":
          description: image dimensions exceed the configured limit
          schema:
            type: string
        "429":
          description: rate limit exceeded
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Upload an image
      tags:
      - Attachment
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT returned by /auth/login.
      Example: "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...